modifyProducts

+++ src/app/products/page.js
@@ -7,7 +7,7 @@ import Edit from "@/components/Edit";
 import Button from "@/components/UI/Button";
 import CardProduct from "@/components/UI/card/product_card";
 import FormUser from "@/components/UI/form/product_form";
-
+import { Toaster, toast } from "react-hot-toast";
 export default function Products() {
   const title = "products";
   const db_name = "products";
@@ -38,11 +38,20 @@ export default function Products() {
   }, [db_name]);
 
   const deleteData = async (id) => {
-    try {
-      await FetchApi({ url: `/api/${db_name}/${id}`, method: "DELETE" });
-      setDataList(dataList.filter((item) => item.id !== id));
-    } catch (error) {
-      console.error("Deletion error: ", error);
+    const confirmDelete = window.confirm(
+      "Êtes-vous sûr de vouloir supprimer cet élément ?"
+    );
+    if (confirmDelete) {
+      try {
+        await FetchApi({ url: `/api/${db_name}/${id}`, method: "DELETE" });
+        setDataList(dataList.filter((item) => item.id !== id));
+        console.log(`Product with ID ${id} deleted successfully`);
+        toast.success(`Product deleted successfully`);
+      } catch (error) {
+        console.error("Deletion error: ", error);
+        console.log(`Failed to delete product with ID ${id}`);
+        toast.error(`Failed to delete product with ID ${id}`);
+      }
     }
   };
 
@@ -59,6 +68,8 @@ export default function Products() {
           dataList={dataList}
         />
       )}
+
+      <Toaster />
 
       <div className={styles.listContainer}>
         <h1 className={styles.listTitle}>{title}</h1>
